% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MRStable}
\alias{MRStable}
\alias{MRStable_V}
\title{Title}
\usage{
MRStable(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  pi_thr1 = 0.6,
  rd = 20,
  over.dispersion = T
)

MRStable_V(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  pi_thr1 = 0.6,
  pi_thr2 = 0.4,
  rd = 20,
  dp = 20,
  over.dispersion = T
)
}
\arguments{
\item{beta_exp}{Exposure}

\item{beta_out}{Outcome}

\item{se_exp}{Std error exposure}

\item{se_out}{Std error outcome}

\item{pi_thr1}{Threshold 1}

\item{rd}{Randomized lasso}

\item{over.dispersion}{Over dispersion}

\item{pi_thr2}{Threshold 2}
}
\value{
A list
}
\description{
Title
}
\examples{
attach(bmi.bmi)
MRStable(beta.exposure, beta.outcome, se.exposure, se.outcome)
}
