% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MRStable}
\alias{MRStable}
\alias{MRStable_V}
\alias{ldsc_divw}
\alias{ada_ldsc_divw}
\title{Title}
\usage{
MRStable(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  pi_thr = 0.6,
  rd = 50,
  over.dispersion = T
)

MRStable_V(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  pi_thr = 0.6,
  rd = 50,
  majority = F,
  over.dispersion.stg1 = T,
  over.dispersion.stg2 = T
)

ldsc_divw(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  scale_exp,
  scale_out,
  over.dispersion = T
)

ada_ldsc_divw(
  beta_exp,
  beta_out,
  se_exp,
  se_out,
  scale_exp,
  scale_out,
  n,
  maxit = 10000
)
}
\arguments{
\item{beta_exp}{Exposure}

\item{beta_out}{Outcome}

\item{se_exp}{Std error exposure}

\item{se_out}{Std error outcome}

\item{pi_thr}{Threshold}

\item{rd}{Randomized lasso}

\item{over.dispersion}{Over dispersion}
}
\value{
A list
}
\description{
Title
}
\examples{
attach(bmi.bmi)
MRStable(beta.exposure, beta.outcome, se.exposure, se.outcome)
}
